IL2CPP code refactoring
* Reorganized the code into different modules for easier maintenance
* Implemented new sigscans for new IL2CPP releases
* Implemented new logic for the type info definition table (thanks to ero for implementing this first on asl-help)
* Some small code optimization
